-Make fonts show up reliably again. Qt4 does this optimization anyway, and it doesn't get confused by NoPen PenStyle.
-Fix testregression build..

svn path=/trunk/KDE/kdelibs/; revision=501728

+++ khtml/test_regression.cpp
@@ -672,7 +672,7 @@ int main(int argc, char *argv[])
     cfg.writeEntry( "Fonts", QStringList() );
     cfg.writeEntry( "DefaultEncoding", "" );
     cfg.setGroup("Java/JavaScript Settings");
-    cfg.writeEntry( "WindowOpenPolicy", KHTMLSettings::KJSWindowOpenAllow);
+    cfg.writeEntry( "WindowOpenPolicy", (int)KHTMLSettings::KJSWindowOpenAllow);
 
     cfg.sync();
